Add new revision properties in SourceStamp
These properties are required to differentiate between requested revision and actual workspace revision.
A build includes both these properties as 'revision' and 'got_revision' but the SourceStamp 'revision' is often 'got_revision'.
In the case where the workspace has a partial checkout, 'got_revision' may not represent the desired revision.

This caused issues with a job creating archives with results from other jobs.
The job only needed a minimal checkout from revision control, with the build configuration,
causing the revision identifier in the web view to stagnate, and builds to not appear in the grid view.