Fix lost slave detection
A build slave was not detected as down if the connection was still active but the slave did not respond.
In particular, if the slave was running a job, the master would detect a job timeout and try to kill the job, but the slave could not respond.

The master now checks for success on keepalive messages, 5 failures or silences cause the slave to be forcefully disconnected.
This in turn will make jobs on the slave killed and retried by the master.

Tested by dropping all packets on the buildbot port.