Fix web grid showing multiple rows for the same revision
The key used for differentiating build rows in the grid was changed in 34fb9c7
to use the revisions for the request and the workspace to properly differentiate builds with a partial workspace.

The key no longer includes the workspace revision so that all builds for the same revision are correctly grouped.
The workspace revision is set during the build, and may not be set at all if the build fails.
This caused the grid to have several rows for the same revision when there were errors or ongoing builds.