Change how got_revision is set for Perforce
The Perforce source step synced using the build request revision and then checked changes on the local workspace to set got_revision.
This caused got_revision to not match the request revision for partial checkout or for revisions only deleting files.

This revision changes got_revision to be a unique revision identifier for repeatable builds.
If the request revision is not set, meaning head, the step sets got_revision to the latest submitted revision.
If the request revision is already unique, a changelist number or a label, the step uses it as got_revision.
Then the source step sync the repository to got_revision.